+++ python/tvm/contrib/hexagon/meta_schedule.py
@@ -21,7 +21,7 @@ import tempfile
 from collections.abc import Callable
 
 import tvm
-from tvm.contrib.popen_pool import PopenPoolExecutor
+from tvm.support.popen_pool import PopenPoolExecutor
 from tvm.driver import build as tvm_build
 from tvm.ir.module import IRModule
 from tvm.runtime import Module, Tensor
